third_party/WebKit/LayoutTests/fast/css/font-family-initial-shorthand-expected.txt

Unstyled Font Family
Initial Font Family
Initial Font

Test for Bug 19085: font-family: initial sets font-family differently than font: initial

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS initialFontFamily is unstyled
PASS initialFont is unstyled
PASS successfullyParsed is true

TEST COMPLETE
